examples/ipywidgets/requirements.txt
--extra-index-url https://pyshiny.netlify.app/pypi/
ipyshiny
